done tinh nang thong bao trong profile

+++ BE/SocialV/src/main/java/com/example/socialv/service/PostCommentService/PostCommentService.java
@@ -2,13 +2,11 @@ package com.example.socialv.service.PostCommentService;
 
 import com.example.socialv.model.Post;
 import com.example.socialv.model.PostComment;
-import com.example.socialv.model.Users;
 import com.example.socialv.repository.ICommentLikeRepository;
 import com.example.socialv.repository.IPostCommentRepository;
 import com.example.socialv.service.NotificationService.INotificationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
@@ -34,7 +32,7 @@ public class PostCommentService implements IPostCommentService{
     @Override
     @Transactional
     public void save(PostComment postComment) {
-        if (postComment.getId() == null){
+        if ((postComment.getId() == null) && (postComment.getPost().getUsers().getId() != postComment.getUsers().getId())){
             notificationService.deleteNotification(postComment.getPost().getId(), 3L);
             notificationService.createNotification(postComment.getUsers().getId(), postComment.getPost().getId(), 3L);
         }
@@ -42,6 +40,7 @@ public class PostCommentService implements IPostCommentService{
     }
 
     @Override
+    @Transactional
     public void remove(Long id) {
         PostComment  comment = findById(id).get();
         commentLikeRepository.deleteAllByComment(comment);
@@ -64,5 +63,10 @@ public class PostCommentService implements IPostCommentService{
             commentLikeRepository.deleteAllByComment(postComment);
         }
         postCommentRepository.deleteAllByPost(post);
+    }
+
+    @Override
+    public Integer countUserCommentPost(Long postId) {
+        return postCommentRepository.countUserCommentPost(postId);
     }
 }
